refactor: move client methods to methods folder
