+++ packages/graz/src/actions/account.ts
@@ -1,5 +1,4 @@
 import type { SigningCosmWasmClientOptions } from "@cosmjs/cosmwasm-stargate";
-import type { Coin } from "@cosmjs/proto-signing";
 import { GasPrice } from "@cosmjs/stargate";
 import type { Key } from "@keplr-wallet/types";
 
@@ -84,31 +83,6 @@ export async function disconnect(clearRecentChain = false): Promise<void> {
     recentChain: clearRecentChain ? null : x.recentChain,
   }));
   return Promise.resolve();
-}
-
-export async function getBalances(bech32Address: string): Promise<Coin[]> {
-  const { activeChain, signingClients } = useGrazStore.getState();
-
-  if (!activeChain || !signingClients) {
-    throw new Error("No connected account detected");
-  }
-
-  const { defaultSigningClient } = useGrazStore.getState();
-  const balances = await Promise.all(
-    activeChain.currencies.map(async (item) => {
-      return signingClients[defaultSigningClient].getBalance(bech32Address, item.coinMinimalDenom);
-    }),
-  );
-
-  return balances;
-}
-
-export async function getStakedBalances(bech32Address: string): Promise<Coin | null> {
-  const { clients } = useGrazStore.getState();
-  if (!clients?.stargate) {
-    throw new Error("Stargate client is not ready");
-  }
-  return clients.stargate.getBalanceStaked(bech32Address);
 }
 
 export function reconnect(): void {
