+++ packages/graz/src/actions/methods.ts
@@ -1,0 +1,28 @@
+import type { Coin } from "@cosmjs/proto-signing";
+
+import { useGrazStore } from "../store";
+
+export async function getBalances(bech32Address: string): Promise<Coin[]> {
+  const { activeChain, signingClients } = useGrazStore.getState();
+
+  if (!activeChain || !signingClients) {
+    throw new Error("No connected account detected");
+  }
+
+  const { defaultSigningClient } = useGrazStore.getState();
+  const balances = await Promise.all(
+    activeChain.currencies.map(async (item) => {
+      return signingClients[defaultSigningClient].getBalance(bech32Address, item.coinMinimalDenom);
+    }),
+  );
+
+  return balances;
+}
+
+export async function getStakedBalances(bech32Address: string): Promise<Coin | null> {
+  const { clients } = useGrazStore.getState();
+  if (!clients?.stargate) {
+    throw new Error("Stargate client is not ready");
+  }
+  return clients.stargate.getBalanceStaked(bech32Address);
+}
